Reworked base tool classes
The BaseToolUI class was doing too much so this has been stripped back to a bare API.
Its logic has been moved into DefaultBaseToolUI along with the camera behaviour.
