+++ amulet_map_editor/programs/edit/plugins/tools/chunk.py
@@ -6,7 +6,7 @@ from OpenGL.GL import (
 )
 
 from amulet_map_editor.api.opengl.camera import Projection
-from amulet_map_editor.programs.edit.api.ui.tool import CameraToolUI
+from amulet_map_editor.programs.edit.api.ui.tool import DefaultBaseToolUI
 from amulet_map_editor.programs.edit.api.behaviour import ChunkSelectionBehaviour
 from amulet.operations.delete_chunk import delete_chunk
 from amulet_map_editor.programs.edit.plugins.operations.stock_plugins.internal_operations.prune_chunks import (
@@ -17,10 +17,10 @@ if TYPE_CHECKING:
     from amulet_map_editor.programs.edit.api.canvas import EditCanvas
 
 
-class ChunkTool(wx.BoxSizer, CameraToolUI):
+class ChunkTool(wx.BoxSizer, DefaultBaseToolUI):
     def __init__(self, canvas: "EditCanvas"):
         wx.BoxSizer.__init__(self, wx.HORIZONTAL)
-        CameraToolUI.__init__(self, canvas)
+        DefaultBaseToolUI.__init__(self, canvas)
 
         self._selection = ChunkSelectionBehaviour(self.canvas)
